Clean: Merge STEX PolygonRasterizer
