+++ src/dewarping/CMakeLists.txt
@@ -20,7 +20,6 @@ SET(
         STEX_GaussBlur.cpp STEX_GaussBlur.h
         STEX_HomographicTransform.h
         STEX_PlugHoles.h
-        STEX_PolygonRasterizer.cpp STEX_PolygonRasterizer.h
         STEX_RasterOpGeneric.h
         STEX_ToPoint.h
         STEX_ToVec.h
